Revert "[refactor] files now in tree order in client"
