+++ app/src/components/App.js
@@ -1,12 +1,9 @@
-import React from 'react'
-import axios from 'axios'
-
-import Chat from './chat/Chat'
-import LoginSignupView from './login/LoginSignupView'
-import ViewNavBar from './topBar/ViewNavbar'
-
-//import Chatroom from './Chatroom'
-//import ChatSelection from './ChatSelection'
+import React from 'react';
+import axios from 'axios';
+import LoginSignupView from './LoginSignupView.js';
+import ViewNavBar from './ViewNavbar.js';
+import Chatroom from './Chatroom.js';
+import ChatSelection from './ChatSelection.js';
 
 
 class App extends React.Component {
@@ -29,76 +26,6 @@ class App extends React.Component {
     this.handleRoomChange = this.handleRoomChange.bind(this);
   }
 
-//===========================================================
-//              Top Bar Methods
-//===========================================================
- handleUserLogout(){
-   var self = this;
-   axios.post('/logout', {id :this.state.userId})
-   .then(res => {
-     self.setState({
-       userId : null,
-       roomId : null,
-       name : null,
-       chat_view : false,
-       login_signup_view : true
-     })
-   })
-   .catch(err => {
-     console.log(err);
-   })
- }
-
-  handleChatExit(){
-   var self = this;
-   if (this.state.roomId) {
-    axios.post('/exitChat', {id : this.state.userId})
-    .then(res => {
-      self.setState({
-        chat_view : false,
-        roomId : null
-      })
-    })
-    .catch(err => {
-      console.log(err);
-    })
-   }
- }
-
-//===========================================================
-//              Login Methods
-//===========================================================
- handleUserSignupLogin(res){
-   this.setState({
-     userId : res.id,
-     name : res.firstname,
-     login_signup_view  : false
-   })
- }
-
-//===========================================================
-//              Chatroom Methods
-//===========================================================
- handleRoomChange(newRoom) {
-   this.setState({
-     roomId : newRoom,
-   })
- }
-
-//===========================================================
-//              ChatSelection Methods
-//===========================================================
- handleChatSelection(inputRoomId, searchOptions, result){
-   this.setState({
-     roomId : inputRoomId,
-     roomSearch : {'option' : searchOptions, 'res' : result},
-     chat_view : true
-   })
- }
-
-//===========================================================
-//              Lifecycle Methods
-//===========================================================
   componentWillMount(){
    axios.get('/checkSession')
    .then(res => {
@@ -131,57 +58,82 @@ class App extends React.Component {
    })
   }
 
+ handleUserSignupLogin(res){
+   this.setState({
+     userId : res.id,
+     name : res.firstname,
+     login_signup_view  : false
+   })
+ }
+
+ handleUserLogout(){
+   var self = this;
+   axios.post('/logout', {id :this.state.userId})
+   .then(res => {
+     self.setState({
+       userId : null,
+       roomId : null,
+       name : null,
+       chat_view : false,
+       login_signup_view : true
+     })
+   })
+   .catch(err => {
+     console.log(err);
+   })
+ }
+
+ handleChatSelection(inputRoomId, searchOptions, result){
+   this.setState({
+     roomId : inputRoomId,
+     roomSearch : {'option' : searchOptions, 'res' : result},
+     chat_view : true
+   })
+ }
+
+ handleChatExit(){
+   var self = this;
+   if (this.state.roomId) {
+    axios.post('/exitChat', {id : this.state.userId})
+    .then(res => {
+      self.setState({
+        chat_view : false,
+        roomId : null
+      })
+    })
+    .catch(err => {
+      console.log(err);
+    })
+   }
+ }
+
+ handleRoomChange(newRoom) {
+   this.setState({
+     roomId : newRoom,
+   })
+ }
+
 
   render() {
-    let navBarProps =  {
-      logout : this.handleUserLogout,
-      home : this.handleChatExit,
-      userId : this.state.userId
-    }
-
-    let chatProps = {
-      roomChange : this.handleRoomChange,
-      userId : this.state.userId,
-      roomId : this.state.roomId, 
-      name : this.state.name,
-      searchResults : this.state.roomSearch,
-      selectRoom : this.handleChatSelection,
-      chat_view :  this.state.chat_view
-    }
-
-    let NavBar = <ViewNavBar {...navBarProps}/>
-    let Login = <LoginSignupView userSignupLogin = {this.handleUserSignupLogin} />
-    let Chat = <Chat {...chatProps}/>
-    
-    if (this.state.mounted) {
-      return (
-        <div>
-          {NavBar}
-          this.state.LoginSignupView ? {Login} : {Chat}
-        </div>
-      )
-    } else {
-      return null;
-    }
-    
-    // if (this.state.mounted){
-    //   if (this.state.login_signup_view) {
-    //     b = (<LoginSignupView userSignupLogin = {this.handleUserSignupLogin}/>)
-    //   } else {
-    //     if (this.state.chat_view) {
-    //       c = (<Chatroom roomChange = {this.handleRoomChange} 
-    //               userId = {this.state.userId} 
-    //               roomId = {this.state.roomId} 
-    //               name = {this.state.name} 
-    //               searchResults = {this.state.roomSearch}/>)
-    //     } else {
-    //       c = (<ChatSelection selectRoom = {this.handleChatSelection}/>)
-    //     }
-    //   }
-    // } else {
-    //   a = (<div></div>)
-    // }
-
+    return (
+      <div>
+        <ViewNavBar logout = {this.handleUserLogout} 
+                    home = {this.handleChatExit} 
+                    userId = {this.state.userId}/>
+       {
+         this.state.mounted ? 
+         (this.state.login_signup_view ? 
+         (<LoginSignupView userSignupLogin = {this.handleUserSignupLogin}/>) :
+         (this.state.chat_view ? <Chatroom roomChange = {this.handleRoomChange} 
+                                           userId = {this.state.userId} 
+                                           roomId = {this.state.roomId} 
+                                           name = {this.state.name} 
+                                           searchResults = {this.state.roomSearch}/> 
+         : < ChatSelection selectRoom = {this.handleChatSelection}/>))  
+         :(<div></div>)
+       }
+      </div>
+    )
   }
 }
