+++ app/src/components/ChatLineItem.js
