+++ app/src/components/Chatroom.js
@@ -1,13 +1,19 @@
 import React, { Component } from 'react';
-import io from 'socket.io-client';
-import axios from 'axios';
-import { Modal, Panel, Alert } from 'react-bootstrap';
-import { Grid, Row, Col, Button } from 'react-bootstrap';
-import { Form, FormGroup, FormControl } from 'react-bootstrap';
-
 import ChatLine from './ChatLineItem';
 import UserItem from './UserItem';
 import ChatJoinModal from './ChatJoinModal.js'
+import io from 'socket.io-client';
+import axios from 'axios';
+import { Modal } from 'react-bootstrap';
+import { Button } from 'react-bootstrap';
+import { Alert } from 'react-bootstrap';
+import { Row } from 'react-bootstrap';
+import { Col } from 'react-bootstrap';
+import { Grid } from 'react-bootstrap';
+import { Panel } from 'react-bootstrap';
+import { Form } from 'react-bootstrap';
+import { FormGroup } from 'react-bootstrap';
+import { FormControl } from 'react-bootstrap';
 
 class Chatroom extends Component {
   constructor(props){
