+++ app/src/components/GoogleMaps/components/InfoWindow.js
