+++ app/src/components/GoogleMaps/components/Marker.js
