+++ app/src/components/GoogleMaps/lib/String.js
