+++ app/src/components/GoogleMaps/lib/cancelablePromise.js
