+++ app/src/components/UserClickInterestsView.js
