+++ app/src/components/ViewNavbar.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
-import { Navbar, Nav, NavItem } from 'react-bootstrap';
-
+import { Navbar } from 'react-bootstrap';
+import { Nav } from 'react-bootstrap';
+import { NavItem } from 'react-bootstrap';
 import UserInterests from './UserInterests'
-import MapModal from './MapModal';
+import MapModal from './MapModal.js';
 
 class ViewNavBar extends Component {
   constructor(props){
