+++ app/src/reducers/sampleReducer.js
@@ -1,0 +1,1 @@
+export default reducer
