+++ package.json
@@ -32,7 +32,6 @@
     "react-bootstrap": "^0.30.7",
     "react-dom": "^15.4.2",
     "react-loading": "0.0.9",
-    "react-redux": "^5.0.2",
     "redux": "^3.6.0",
     "redux-logger": "^2.8.0",
     "redux-promise-middleware": "^4.2.0",
@@ -45,7 +44,6 @@
     "babel-core": "^6.21.0",
     "babel-loader": "^6.2.10",
     "babel-plugin-transform-class-properties": "^6.19.0",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-object-rest-spread": "^6.22.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
@@ -57,12 +55,12 @@
     "lodash": "^4.17.4",
     "mocha": "^3.2.0",
     "morgan": "^1.7.0",
-    "mysql": "^1.1.2",
     "react-hot-loader": "^1.3.0",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
     "validator": "^6.2.1",
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.12.1"
+    "webpack-dev-server": "^1.12.1",
+    "mysql": "^1.1.2",
+    "webpack": "^1.14.0"
   }
 }
